i3blocks/pyweather/pyproject.toml
[project]
name = "pyweather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "pyinstaller>=6.17.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
